# docs/swagger.yaml
basePath: /
definitions:
  internal_handler.LoginRequest:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
  internal_handler.LoginResponse:
    properties:
      token:
        type: string
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: "123456"
        type: string
    type: object
  wishlist-api_internal_domain.Item:
    properties:
      description:
        type: string
      id:
        type: integer
      is_booked:
        type: boolean
      link:
        type: string
      priority:
        type: integer
      title:
        type: string
      wishlist_id:
        type: integer
    type: object
  wishlist-api_internal_domain.Wishlist:
    properties:
      description:
        type: string
      event_date:
        type: string
      id:
        type: integer
      short_token:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис для создания списков желаний.
  title: Wishlist API
  version: "1.0"
paths:
  /api/wishlists:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/wishlist-api_internal_domain.Wishlist'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Мои вишлисты
      tags:
      - private
    post:
      consumes:
      - application/json
      parameters:
      - description: Параметры списка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/wishlist-api_internal_domain.Wishlist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wishlist-api_internal_domain.Wishlist'
      security:
      - ApiKeyAuth: []
      summary: Создать вишлист
      tags:
      - private
  /api/wishlists/{id}:
    delete:
      parameters:
      - description: ID списка
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Удалить вишлист
      tags:
      - private
  /api/wishlists/{id}/items:
    post:
      parameters:
      - description: ID вишлиста
        in: path
        name: id
        required: true
        type: integer
      - description: Данные подарка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/wishlist-api_internal_domain.Item'
      responses:
        "201":
          description: Created
      security:
      - ApiKeyAuth: []
      summary: Добавить подарок
      tags:
      - private
  /api/wishlists/items/{itemId}:
    delete:
      parameters:
      - description: ID подарка
        in: path
        name: itemId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Удалить подарок
      tags:
      - private
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.LoginResponse'
      summary: Логин
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Регистрация
      tags:
      - auth
  /share/{token}:
    get:
      parameters:
      - description: Токен вишлиста
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Публичный просмотр
      tags:
      - public
  /share/{token}/book/{itemID}:
    post:
      parameters:
      - description: Токен вишлиста
        in: path
        name: token
        required: true
        type: string
      - description: ID подарка
        in: path
        name: itemID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Already booked
          schema:
            type: string
      summary: Бронирование подарка
      tags:
      - public
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"